refactor: CLLocationManagerDelegate를 ViewController에서 분리하기
